requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
boto3==1.9.66
botocore==1.12.66         # via boto3, s3transfer
dnspython==1.16.0
docopt==0.6.2
docutils==0.14            # via botocore
futures==3.2.0            # via s3transfer
jmespath==0.9.3           # via boto3, botocore
numpy==1.15.4
pyserial==3.4
python-dateutil==2.7.5    # via botocore
pytz==2018.7
rpi.gpio==0.6.5
s3transfer==0.1.13        # via boto3
six==1.12.0               # via python-dateutil
urllib3==1.24.1           # via botocore
wheel==0.32.3
